dom/base/test/chrome/file_bug1209621.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1209621
-->
<window title="Mozilla Bug 1209621"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
  onload="run()">
  <label value="Mozilla Bug 1209621"/>
  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[
  function ok(cond, msg) {
    opener.wrappedJSObject.ok(cond, msg);
  }

  function is(actual, expected, msg) {
    opener.wrappedJSObject.is(actual, expected, msg);
  }

  function run() {
    var docshell = window.docShell;
    ok(docshell, "Active window should have a DocShell");
    var treeOwner = docshell.treeOwner;
    ok(treeOwner, "Active docshell should have a TreeOwner!");

    is(treeOwner.primaryContentShell, null,
       "There shouldn't be primaryContentShell because no browser has primary=true.");
    is(treeOwner.primaryRemoteTab, null,
       "There shouldn't be primaryRemoteTab because no remote browser has primary=true.");

    var ip = document.getElementById("inprocess");
    var remote = document.getElementById("remote");
    var remote2 = document.getElementById("remote2");

    ip.setAttribute("primary", "true");
    ok(ip.docShell, "non-remote browser should have a DocShell.");
    is(treeOwner.primaryContentShell, ip.docShell,
       "primary browser should be the primaryContentShell.");
    is(treeOwner.primaryRemoteTab, null,
       "There shouldn't be primaryRemoteTab because no remote browser has primary=true.");

    ip.removeAttribute("primary");
    remote.setAttribute("primary", "true");
    is(treeOwner.primaryContentShell, null,
       "There shouldn't be primaryContentShell because no browser has primary=true.");
    var tp = remote.frameLoader.remoteTab;
    ok(tp, "Remote browsers should have a remoteTab.");
    is(treeOwner.primaryRemoteTab, tp,
       "primary remote browser should be the primaryRemoteTab.");

    remote.removeAttribute("primary");
    is(treeOwner.primaryContentShell, null,
       "There shouldn't be primaryContentShell because no browser has primary=true.");
    is(treeOwner.primaryRemoteTab, null,
       "There shouldn't be primaryRemoteTab because no remote browser has primary=true.");

    remote2.setAttribute("primary", "true");
    var tp2 = remote2.frameLoader.remoteTab;
    ok(tp2, "Remote browsers should have a remoteTab.");
    is(treeOwner.primaryRemoteTab, tp2,
       "primary remote browser should be the primaryRemoteTab.");
    is(treeOwner.primaryContentShell, null,
       "There shouldn't be primaryContentShell because no browser has primary=true.");

    opener.setTimeout(function() { this.done(); }, 0);
    window.close();
  }

  ]]></script>
  <browser type="content" src="about:blank" id="inprocess"/>
  <browser type="content" remote="true" src="about:blank" id="remote"/>
  <browser type="content" remote="true" src="about:blank" id="remote2"/>
</window>
